+++ test/transport/websocket/websocket_transport_html_server.dart
@@ -1,7 +1,10 @@
+@TestOn('browser')
+
 import 'dart:io';
 
 import 'package:connectanum/src/message/message_types.dart';
 import 'package:stream_channel/stream_channel.dart';
+import 'package:test/test.dart';
 
 // Once the hybrid isolate starts, it will call the special function
 // hybridMain() with a StreamChannel that's connected to the channel
@@ -12,9 +15,11 @@ hybridMain(StreamChannel channel) async {
     print("receive open request");
     if (req.uri.path == '/wamp') {
       var socket = await WebSocketTransformer.upgrade(req);
+      print("Received protocol" + socket.protocol);
       socket.listen((message) {
         print("Received Message: " + message);
-        if (message is String && message.contains("["+ MessageTypes.CODE_HELLO.toString())) {
+        if (message is String &&
+            message.contains("[" + MessageTypes.CODE_HELLO.toString())) {
           socket.add("[" + MessageTypes.CODE_WELCOME.toString() + ",1234,{}]");
         }
       });