+++ test/transport/websocket/websocket_transport_io_test.dart
@@ -12,15 +12,19 @@ import 'package:test/test.dart';
 
 void main() {
   group('WebSocket protocol with io communication', () {
-    test('Opening a server connection and simple send receive scenario using a serializer', () async {
+    test(
+        'Opening a server connection and simple send receive scenario using a serializer',
+        () async {
       var server = await HttpServer.bind('localhost', 9100);
       var serializer = Serializer();
       server.listen((HttpRequest req) async {
         if (req.uri.path == '/wamp') {
           var socket = await WebSocketTransformer.upgrade(req);
           socket.listen((message) {
-            if (message is String && message.contains("["+ MessageTypes.CODE_HELLO.toString())) {
-              socket.add("[" + MessageTypes.CODE_WELCOME.toString() + ",1234,{}]");
+            if (message is String &&
+                message.contains("[" + MessageTypes.CODE_HELLO.toString())) {
+              socket.add(
+                  "[" + MessageTypes.CODE_WELCOME.toString() + ",1234,{}]");
             }
           });
         }
@@ -29,8 +33,7 @@ void main() {
       WebSocketTransport transport = WebSocketTransport(
           "ws://localhost:9100/wamp",
           serializer,
-          WebSocketTransport.SERIALIZATION_JSON
-      );
+          WebSocketTransport.SERIALIZATION_JSON);
 
       await transport.open();
       transport.send(Hello("my.realm", Details()));