+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 1.0.0-dev.6
+
+- fixed hello.details serialization code
+- added error in serializer for incoming messages
+- added serializer logging for wrong incoming messages
+- more cleanup code to meet pana analysis requirements
+
 ### 1.0.0-dev.5
 
 - more cleanup code to meet pana analysis requirements